Add --require-barcodes-both-ends CLI option
Add a new CLI option --require-barcodes-both-end which requires that a
barcode match is found at both ends of the read in order to be
accepted when deplexing. Reads with a match at only one end are
classified as unassigned.

A basic test for BarcodeScannerEPI2ME using the PBC096 kit is
included. The other scanners are not tested explicitly.